I've updated the text element rendering to add editing capabilities. Here are the changes I made:
Added contentEditable attribute to make the text element editable
Added suppressContentEditableWarning to prevent React warnings about contentEditable
Added onBlur handler to save changes when user clicks away
Added onKeyDown handler to save changes when user presses Enter
The text content is updated in the store using updateElement when changes are made
The text elements should now be editable with the following workflow:

Click on a text element to select it
Click again to start editing
Type your changes
Press Enter or click outside to save changes
The text will remain draggable using the edges of the element (when selected), and you can still resize it using the resize handles. The editing mode is activated only when you click directly on the text.

Try adding a new text element and you should now be able to:

Drag it around using the edges
Resize it using the handles when selected
Edit the text by clicking on it
Save changes by pressing Enter or clicking outside
Let me know if you'd like any adjustments to this behavior or if you encounter any issues!